calc_order: "series" to "seriesvar"
In MATLAB version 2015b, on a new computer, I installed all of the
toolboxes that come with the OSU matlab license. There was a conflict
because this function uses the variable “series” and that is the name
of a function in one of the toolboxes.

Changed all instances of “series” to “seriesvar”. No functional changes
made.