added comment to get_chipod_cals
Time steps and averaging are computed in different ways for older chi
pods with 120Hz T1P and T2p versus the newer chi pods with 100Hz T1P
and T2P. At first I thought the code was incorrect for the newer chi
pods, but it is actually okay. Added comment to stress this.