changes to get_chipod_raw
For the newer 100 Hz chipods, a number of extraneous fields are saved
at weird intervals, sometimes not at regular frequencies. This code
ensures that any variable not saved at correct frequencies is not
included in the structure ‘data’.

Also added a new section for chipods with header version 80 (100Hz
chipods). It is almost identical to the code in the else statement that
starts on line 209.