updated raw_load_chipod for OLD chipod files and NEW matlab
trying to reprocess chipod files from 2011 with Matlab2016a. These
chipod files have header version 48. The names of each variable are
padded with lots of spaces so they are all length=12. But in this newer
version of Matlab, names with spaces cannot be used to create variables
with a structure. This had been fixed for newer header versions, but
header version 48 had not been updated to get rid of the extra spaces
in the names.